requirements.txt
# Core ML/Data Science
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
xgboost==2.1.2
lightgbm==4.5.0

# MLOps
mlflow==2.16.2
dvc==3.56.0
dvc-s3==3.2.0

# API Framework
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
alembic==1.13.3

# Monitoring
prometheus-client==0.21.0
evidently==0.4.33

# AWS Integration
boto3==1.35.0
awscli==1.34.0

# Data Formats
pyarrow==17.0.0
openpyxl==3.1.5

# Configuration
pyyaml==6.0.2
python-dotenv==1.0.1

# Utilities
joblib==1.4.2
requests==2.32.3
click==8.1.7

# Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0
httpx==0.27.2

# Code Quality
black==24.10.0
ruff==0.7.1
isort==5.13.2
mypy==1.13.0

# Visualization (for reports)
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1

# Logging
loguru==0.7.2
